SPIDEV compatibility layer improvement and patch for 64bits (#455)
* SPIDEV GPIO exception and optimization(via caching sysfs)
#407

* Fix spidev compatibility layer for 64bit systems + monotic clock

+++ utility/SPIDEV/compatibility.h
@@ -3,6 +3,7 @@
  * Author: purinda
  *
  * Created on 24 June 2012, 3:08 PM
+ * patch for safer monotonic clock & millis() correction for 64bit LDV 2018
  */
 
 #ifndef COMPATIBLITY_H
@@ -11,7 +12,8 @@
 #ifdef	__cplusplus
 extern "C" {
 #endif
-	
+
+#include <stdint.h>  // for uintXX_t types	
 #include <stddef.h>
 #include <time.h>
 #include <sys/time.h>
@@ -19,7 +21,7 @@ extern "C" {
 void __msleep(int milisec);
 void __usleep(int milisec);
 void __start_timer();
-long __millis();
+uint32_t __millis();
 
 #ifdef	__cplusplus
 }
